Creating about and home component, moved app.jsx content to home component. Importing Home, Header and About to app.jsx and adding routes and switch.
