+++ src/components/About.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Container } from 'semantic-ui-react';
+
+const About = () => {
+  return (
+    <Container>
+      
+    </Container>
+  )
+}
+
+export default About;
