+++ src/components/DisplayPerformanceData.jsx
@@ -60,7 +60,7 @@ class DisplayPerformanceData extends Component {
           'rgba(255, 206, 86, 0.2)',
           'rgba(75, 192, 192, 0.2)'
         ],
-        
+      
       }],
     };
